+++ client/changelog.md
@@ -1,1 +1,2 @@
 ### Client
+- Fixed crackling/noise in microphone capture on Minecraft 26.1.x.